refactor: landygraphfactory
